Add runtime structures for the VM and add a link step to create them
We link CFUNCTIONs and CMODULEs into real bytecode objects and
interpret those instead. This allows concatenating bytecode vectors
for functions into bytecode vectors for modules.

Also change a bunch of VM stuff with stack and base pointer
manipulation to use frame size. This will probably need to get tweaked
when real argument processing becomes a thingg. (And when we actually
figure out what to do with RETURN and RETURN-FROM with respect to mv
handling.)