Unify/simplify/speed-up non-local exit handling.
* Instead of having distinct instructions for
tagbody/go/block/return-from, use a unified set of instructions for
handling both. This simplifies and speeds up the VM, as now the
targets are directly encoded at the time of control transfer, and no
indirection is needed. The specification doc has more info about this
set of primitives and the advantages.
* The bytecode compiler has been updated to handle this. Notice that
the handling is simpler since there's no indirection. Support for
blocks has not been written yet but this change should make it just as
easy if not easier to implement on the compiler side.