Fix BIND reversal, again
When we bind (let ((x ...) (y ...)) ...) left to write, we first
push the value of x, then push the value of y, then bind. If bind
copies into the locals without reversing the polarity, as this
change makes it do, we can perform the obvious location assignments
of x = 0, y = 1. It also means a VM implementation can just do a
simple copy to implement bind.