Update
I have been working on the Clasp-particular version of this code
in the Clasp repo. This commit ports most of the changes I've made
to the compiler there. In particular:

* ENTRY et al. are only generated if needed
* Handling of specials in lambda lists corrected
* Infrastructure (related to LOAD-TIME-VALUE) needed for file
  compilation semantics
* Organization
* Various fixes
* Probably some other stuff

Plus it adds an ASDF system. Originally I wanted to break it into
multiple systems: one for the machine definition, one for the
compiler, and one for the VM, but ASDF wasn't cooperating.