Use only one vcell & version bump
Earlier I had it so that symbol-value etc took two arguments, the
symbol and the global cell, but realistically there's no way that
the same symbol would have different global cells in the same code-
so that you locally bind in the same place for two distinct
globals- that's perverse. If necessary it should be possible to
pack both global and local info into the one vcell.

Also bumped the version since vcells+fcells are an incompatibility.