Implement UNWIND-PROTECT
That's all the standard special operators.

An alternative design would be to inline the cleanup, and have the
PROTECT instruction refer to a label. The cleanup would then end
with a special resume instruction, that would either continue
unwinding, or jump back to whatever local exit caused the cleanup.
But that would be really complicated.

Originally I was going to call the cleanup instruction "deprotect"
as a chemistry joke, but I figure deprotect sounds like it's
merely removing protection, rather than actually doing things.