Reorganize
With FASL loading available it makes sense to separate the
compiler and loader components more strictly. This means you can
load cvm/load and a VM and load bytecode, without having to load
in the compiler and trucler and stuff.

Having a million subsystems is obviously a little confusing so I
tried to document it clearly.