github automation: run tests non-interactively
So if they die they die fast rather than hang waiting for input
that will never come.

We use --non-interactive instead of --script because we need the
init file to load quicklisp, and given that the machine is fresh
it's not like there's any other init to worry about skipping.