Maintain debug names and lambda lists for bytecode functions
Besides being nice, this lets us test that the information is
stored, and serialized in FASLs.

Those tests still need to be written.

Also TODO is letting functions from DEFUN, DEFMACRO etc. get names.
In Clasp we used a declaration for this, which I think is the way
to go since it allows other form processors to ignore it, unlike
SBCL's NAMED-LAMBDA or the like. On the other hand there should be
a way to make it client customizable.