Change PROTECT instruction to take a constant parameter
This makes things a little more complex to describe and adds a
little more code. On the other hand, it saves a stack push and pop,
and makes analysis simpler: it's impossible for PROTECT to refer
to some random alien thunk. Also, hypothetically a VM impl could
DX allocate the thunk even before analysis.